TechnicalIndicatorCharts

Documentation for TechnicalIndicatorCharts.

TechnicalIndicatorCharts.abbrev

TechnicalIndicatorCharts.chart

TechnicalIndicatorCharts.extract_value

TechnicalIndicatorCharts.indicator_fields

TechnicalIndicatorCharts.indicator_fields_values

TechnicalIndicatorCharts.ismultiinput

TechnicalIndicatorCharts.ismultioutput

TechnicalIndicatorCharts.make_panel

TechnicalIndicatorCharts.make_panel

TechnicalIndicatorCharts.merge_candle!

TechnicalIndicatorCharts.push_new_candle!

TechnicalIndicatorCharts.random_candles

TechnicalIndicatorCharts.update!

TechnicalIndicatorCharts.update_last_candle!

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.abbrev — Method

abbrev(p::Period)

Return an abbreviated string representation of the given period.

Example

abbrev(Hour(4)) # "4h"
abbrev(Day(1))  # "1d"

TechnicalIndicatorCharts.chart — Method

  chart(name, tf; indicators, visuals)

Construct a Chart instance configured with the given indicators and visual parameters.

Example

julia> golden_cross = chart(
    "BTCUSD", Hour(4);
    indicators = [
        SMA{Float64}(;period=50),
        SMA{Float64}(;period=200)
    ],
    visuals = [
        Dict(
            :label_name => "SMA 50",
            :line_color => "#E072A4",
            :line_width => 2
        ),
        Dict(
            :label_name => "SMA 200",
            :line_color => "#3D3B8E",
            :line_width => 5
        )
    ]
)

TechnicalIndicatorCharts.extract_value — Method

extract_value(value)

Extract values out of an indicators value struct. This is only intended to be used for indicators that emit multiple values per tick.

TechnicalIndicatorCharts.indicator_fields — Method

Return a tuple of symbol names to be used for the output of ind.

TechnicalIndicatorCharts.indicator_fields_values — Method

indicator_fields_values(ind)

Extract values from an indicator instance.

TechnicalIndicatorCharts.ismultiinput — Method

This is a wrapper around OnlineTechnicalIndicators.ismultiinput that takes any instance of a TechnicalIndicator and digs out its unparametrized type before running the original ismultiinput method.

TechnicalIndicatorCharts.ismultioutput — Method

This is a wrapper around OnlineTechnicalIndicators.ismultioutput that takes any instance of a TechnicalIndicator and digs out its unparametrized type before running the original ismultioutput method.

TechnicalIndicatorCharts.make_panel — Method

make_panel(
    chart::LightweightCharts.LWCChart
) -> LightweightCharts.LWCPanel

Wrap a single LWCChart in a panel.

TechnicalIndicatorCharts.make_panel — Method

make_panel(plots::Vector) -> LightweightCharts.LWCPanel

Wrap a Vector of LWCCharts in a panel.

TechnicalIndicatorCharts.merge_candle! — Method

merge_candle!(last_candle, c)

If last candle is not provided, construct a new candle with the given OHLCV data. If last candle is provided, mutate last_candle such that it's HLCV are updated. When tw candles are passed in, it's assumed they have the same timestamp.

TechnicalIndicatorCharts.push_new_candle! — Method

push_new_candle!(chart, c)

This is meant to be called on timeframe boundaries to onto the chart's dataframe. It also does indicator calculation at this time.

TechnicalIndicatorCharts.random_candles — Method

random_candles(tf, n; start)

This belongs somewhere in ~/hak/trading, but it can live here for now.

TechnicalIndicatorCharts.update! — Method

update!(chart, c)

Update a chart with a candle.

TechnicalIndicatorCharts.update_last_candle! — Method

update_last_candle!(chart, c)

This updates the HLCV values of the last row of the chart's DataFrame when we're not at a chart.tf boundary.

TechnicalIndicatorCharts.visualize — Method

visualize(
    unimplemented,
    opts,
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

This is a visualize method that's a catch-all for indicators that haven't had a visualize method made for them yet. For now, it returns missing.

TechnicalIndicatorCharts.visualize — Method

visualize(
    chart::Chart;
    min_height,
    mode
) -> LightweightCharts.LWCLayout

Return an LWCLayout that visualizes all the components in chart appropriately.

TechnicalIndicatorCharts.visualize — Method

visualize(
    df::DataFrames.DataFrame,
    opts
) -> LightweightCharts.LWCChart

Visualize a DataFrame using lwc_candlestick.

TechnicalIndicatorCharts.visualize — Method

visualize(
    bb::OnlineTechnicalIndicators.BB,
    opts,
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize Bollinger Bands using 3 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    ema::OnlineTechnicalIndicators.EMA,
    opts,
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize EMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    hma::OnlineTechnicalIndicators.HMA,
    opts,
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize HMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    rsi::OnlineTechnicalIndicators.RSI,
    opts,
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize RSI using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    sma::OnlineTechnicalIndicators.SMA,
    opts,
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Return an lwc_line for visualizing an SMA indicator.

TechnicalIndicatorCharts.visualize — Method

visualize(
    srsi::OnlineTechnicalIndicators.StochRSI,
    opts,
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize StochRSI using 2 lwc_lines.