TechnicalIndicatorCharts

Documentation for TechnicalIndicatorCharts.

TechnicalIndicatorCharts.Chart

TechnicalIndicatorCharts.abbrev

TechnicalIndicatorCharts.config

TechnicalIndicatorCharts.df_fields

TechnicalIndicatorCharts.extract_value

TechnicalIndicatorCharts.indicator_fields

TechnicalIndicatorCharts.indicator_fields_values

TechnicalIndicatorCharts.ismultiinput

TechnicalIndicatorCharts.ismultioutput

TechnicalIndicatorCharts.make_panel

TechnicalIndicatorCharts.make_panel

TechnicalIndicatorCharts.merge_candle!

TechnicalIndicatorCharts.push_new_candle!

TechnicalIndicatorCharts.update!

TechnicalIndicatorCharts.update_last_candle!

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.Chart — Type

Summary

A Chart is a mutable struct that has:

a name (typically of the asset like "BTCUSD" or "AAPL")

a timeframe (which controls how much time each candle on the chart represents)

a DataFrame to hold OHLCV values and indicator values

a Vector of OnlineTechnicalIndicators to display on the chart.

another Vector of display configuration for each indicator.

If you were to go to TradingView and look at a chart, imagine what kind of data structure would be required to represent it in memory. That's what the Chart struct aims to be.

Fields

name::AbstractString

tf::Dates.Period

indicators::AbstractVector{OnlineTechnicalIndicators.TechnicalIndicator}

visuals::AbstractVector

df::DataFrames.DataFrame

ts::Union{Missing, Dates.DateTime}

candle::Union{Missing, Candle}

Constructors

Chart(name, tf; indicators=[], visuals=[])

Examples

julia> just_candles = Chart("ETHUSD", Minute(1))

julia> golden_cross = Chart(
           "BTCUSD", Hour(4);
           indicators = [
               SMA{Float64}(;period=50),
               SMA{Float64}(;period=200)
           ],
           visuals = [
               Dict(
                   :label_name => "SMA 50",
                   :line_color => "#E072A4",
                   :line_width => 2
               ),
               Dict(
                   :label_name => "SMA 200",
                   :line_color => "#3D3B8E",
                   :line_width => 5
               )
           ]
       )

julia> default_visuals = Chart(
           "BNBUSDT", Hour(4);
           indicators = [BB{Float64}(), StochRSI{Float64}()],
           visuals    = [nothing,       nothing]  # To use defaults, pass in `nothing`.
       )

TechnicalIndicatorCharts.abbrev — Method

abbrev(p::Period)

Return an abbreviated string representation of the given period.

Example

abbrev(Hour(4)) # "4h"
abbrev(Day(1))  # "1d"

TechnicalIndicatorCharts.config — Method

config(sma::OnlineTechnicalIndicators.SMA) -> Vector{Any}

Return the default visualization config for SMA.

TechnicalIndicatorCharts.df_fields — Method

df_fields(indicators) -> Any

Return a list of fields to be used in a DataFrame for chart values.

TechnicalIndicatorCharts.extract_value — Method

extract_value(value) -> Vector{Any}

Extract values out of an indicators value struct. This is only intended to be used for indicators that emit multiple values per tick.

TechnicalIndicatorCharts.indicator_fields — Method

Return a tuple of symbol names to be used for the output of ind.

TechnicalIndicatorCharts.indicator_fields_values — Method

indicator_fields_values(
    ind::OnlineTechnicalIndicators.TechnicalIndicatorMultiOutput
) -> Any

Extract values from an indicator instance.

TechnicalIndicatorCharts.ismultiinput — Method

This is a wrapper around OnlineTechnicalIndicators.ismultiinput that takes any instance of a TechnicalIndicator and digs out its unparametrized type before running the original ismultiinput method.

TechnicalIndicatorCharts.ismultioutput — Method

This is a wrapper around OnlineTechnicalIndicators.ismultioutput that takes any instance of a TechnicalIndicator and digs out its unparametrized type before running the original ismultioutput method.

TechnicalIndicatorCharts.make_panel — Method

make_panel(plots::AbstractVector; h)

Wrap a Vector of LWCCharts in a panel.

TechnicalIndicatorCharts.make_panel — Method

make_panel(chart::LightweightCharts.LWCChart; h)

Wrap a single LWCChart in a panel.

TechnicalIndicatorCharts.merge_candle! — Method

merge_candle!(
    last_candle::Union{Missing, Candle},
    c::Union{Candle, DataFrames.DataFrameRow}
) -> Candle

If last candle is not provided, construct a new candle with the given OHLCV data. If last candle is provided, mutate last_candle such that it's HLCV are updated. When tw candles are passed in, it's assumed they have the same timestamp.

TechnicalIndicatorCharts.push_new_candle! — Method

push_new_candle!(
    chart::Chart,
    c::Candle
) -> DataFrames.DataFrame

This is meant to be called on timeframe boundaries to onto the chart's dataframe. It also does indicator calculation at this time.

TechnicalIndicatorCharts.update! — Method

update!(chart::Chart, c::Candle) -> Union{Nothing, Candle}

Update a chart with a candle. When a candle is completed, return it. Otherwise, return nothing on update.

TechnicalIndicatorCharts.update_last_candle! — Method

update_last_candle!(chart::Chart, c::Candle) -> Float64

This updates the HLCV values of the last row of the chart's DataFrame when we're not at a chart.tf boundary.

TechnicalIndicatorCharts.visualize — Method

visualize(
    unimplemented,
    opts,
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

This is a visualize method that's a catch-all for indicators that haven't had a visualize method made for them yet. For now, it returns missing.

TechnicalIndicatorCharts.visualize — Method

visualize(
    chart::Chart;
    min_height,
    mode,
    up_color,
    down_color
) -> LightweightCharts.LWCLayout

Return an LWCLayout that visualizes all the components in chart appropriately.

TechnicalIndicatorCharts.visualize — Method

visualize(
    df::DataFrames.DataFrame,
    opts
) -> LightweightCharts.LWCChart

Visualize a DataFrame using lwc_candlestick.

TechnicalIndicatorCharts.visualize — Method

visualize(
    adx::OnlineTechnicalIndicators.ADX,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize ADX using 3 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    alma::OnlineTechnicalIndicators.ALMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize ALMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    ao::OnlineTechnicalIndicators.AO,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize AO (Awesome Oscillator) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    atr::OnlineTechnicalIndicators.ATR,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize ATR using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    ad::OnlineTechnicalIndicators.AccuDist,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize AccuDist (Accumulation/Distribution) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    aroon::OnlineTechnicalIndicators.Aroon,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize Aroon using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    bb::OnlineTechnicalIndicators.BB,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize Bollinger Bands using 3 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    bop::OnlineTechnicalIndicators.BOP,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize BOP (Balance of Power) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    cci::OnlineTechnicalIndicators.CCI,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize CCI (Commodity Channel Index) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    chop::OnlineTechnicalIndicators.CHOP,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize CHOP (Choppiness Index) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    co::OnlineTechnicalIndicators.ChaikinOsc,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize ChaikinOsc (Chaikin Oscillator) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    cks::OnlineTechnicalIndicators.ChandeKrollStop,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize ChandeKrollStop using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    cc::OnlineTechnicalIndicators.CoppockCurve,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize CoppockCurve using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    dema::OnlineTechnicalIndicators.DEMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize DEMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    dpo::OnlineTechnicalIndicators.DPO,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize DPO using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    dc::OnlineTechnicalIndicators.DonchianChannels,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize Donchian Channels using 3 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    ema::OnlineTechnicalIndicators.EMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize EMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    emv::OnlineTechnicalIndicators.EMV,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize EMV (Ease of Movement) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    fi::OnlineTechnicalIndicators.ForceIndex,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize ForceIndex using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    hma::OnlineTechnicalIndicators.HMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize HMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    kama::OnlineTechnicalIndicators.KAMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize KAMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    kst::OnlineTechnicalIndicators.KST,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize KST using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    kvo::OnlineTechnicalIndicators.KVO,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize KVO (Klinger Volume Oscillator) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    keltnerchannels::OnlineTechnicalIndicators.KeltnerChannels,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize Keltner Channels (KC) using 3 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    macd::OnlineTechnicalIndicators.MACD,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize MACD using 2 lwc_lines and 1 histogram.

TechnicalIndicatorCharts.visualize — Method

visualize(
    mi::OnlineTechnicalIndicators.MassIndex,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize MassIndex using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    mcginleydynamic::OnlineTechnicalIndicators.McGinleyDynamic,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize McGinleyDynamic using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    md::OnlineTechnicalIndicators.MeanDev,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize MeanDev using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    obv::OnlineTechnicalIndicators.OBV,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize On Balance Volume (OBV) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    psar::OnlineTechnicalIndicators.ParabolicSAR,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize Parabolic SAR using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    phl::OnlineTechnicalIndicators.PivotsHL,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize PivotsHL using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    roc::OnlineTechnicalIndicators.ROC,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize ROC using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    rsi::OnlineTechnicalIndicators.RSI,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize RSI using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    sma::OnlineTechnicalIndicators.SMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Return an lwc_line for visualizing an SMA indicator.

TechnicalIndicatorCharts.visualize — Method

visualize(
    SMMA::OnlineTechnicalIndicators.SMMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Return an lwc_line for visualizing an SMMA indicator.

TechnicalIndicatorCharts.visualize — Method

visualize(
    sobv::OnlineTechnicalIndicators.SOBV,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize SOBV (Smoothed On Balance Volume) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    stc::OnlineTechnicalIndicators.STC,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize STC using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    sd::OnlineTechnicalIndicators.StdDev,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize StdDev using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    stoch::OnlineTechnicalIndicators.Stoch,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize Stochastic Oscillator using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    srsi::OnlineTechnicalIndicators.StochRSI,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize StochRSI using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    st::OnlineTechnicalIndicators.SuperTrend,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize SuperTrend using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    t3::OnlineTechnicalIndicators.T3,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize T3 using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    tema::OnlineTechnicalIndicators.TEMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize TEMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    trix::OnlineTechnicalIndicators.TRIX,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize TRIX using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    tsi::OnlineTechnicalIndicators.TSI,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize True Strength Index (TSI) using 1 lwc_line. Note that on TradingView, TSI includes a second signal line that is not included here.

TechnicalIndicatorCharts.visualize — Method

visualize(
    ttm::OnlineTechnicalIndicators.TTM,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize TTM Squeeze using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    uo::OnlineTechnicalIndicators.UO,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize UO (Ultimate Oscillator) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    vtx::OnlineTechnicalIndicators.VTX,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> Vector{LightweightCharts.LWCChart}

Visualize VTX (Vortex Indicator) using 2 lwc_lines.

TechnicalIndicatorCharts.visualize — Method

visualize(
    vwap::OnlineTechnicalIndicators.VWAP,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize VWAP (Volume Weighted Average Price) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    vwma::OnlineTechnicalIndicators.VWMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize VWMA (Volume Weighted Moving Average) using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    wma::OnlineTechnicalIndicators.WMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize WMA using 1 lwc_line.

TechnicalIndicatorCharts.visualize — Method

visualize(
    zlema::OnlineTechnicalIndicators.ZLEMA,
    opts::Union{Nothing, AbstractDict},
    df::DataFrames.DataFrame
) -> LightweightCharts.LWCChart

Visualize ZLEMA using 1 lwc_line.