TechnicalIndicatorCharts

Documentation for TechnicalIndicatorCharts.

TechnicalIndicatorCharts.abbrev

TechnicalIndicatorCharts.chart

TechnicalIndicatorCharts.indicator_fields

TechnicalIndicatorCharts.indicator_fields_values

TechnicalIndicatorCharts.ismultiinput

TechnicalIndicatorCharts.ismultioutput

TechnicalIndicatorCharts.make_panel

TechnicalIndicatorCharts.make_panel

TechnicalIndicatorCharts.merge_candle!

TechnicalIndicatorCharts.push_new_candle!

TechnicalIndicatorCharts.update_last_candle!

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.visualize

TechnicalIndicatorCharts.abbrev — Method

abbrev(p::Period)

Return an abbreviated string representation of the given period.

Example

abbrev(Hour(4)) # "4h"
abbrev(Day(1))  # "1d"

TechnicalIndicatorCharts.chart — Method

  chart(name, tf; indicators, visuals)

Construct a Chart instance configured with the given indicators and visual parameters.

Example

julia> golden_cross = chart(
    "BTCUSD", Hour(4);
    indicators = [
        SMA{Float64}(;period=50),
        SMA{Float64}(;period=200)
    ],
    visuals = [
        Dict(
            :label_name => "SMA 50",
            :line_color => "#E072A4",
            :line_width => 2
        ),
        Dict(
            :label_name => "SMA 200",
            :line_color => "#3D3B8E",
            :line_width => 5
        )
    ]
)

TechnicalIndicatorCharts.indicator_fields — Method

Return a tuple of symbol names to be used for the output of ind.

TechnicalIndicatorCharts.indicator_fields_values — Method

Extract values from an indicator instance.

TechnicalIndicatorCharts.ismultiinput — Method

This is a wrapper around OnlineTechnicalIndicators.ismultiinput that takes any instance of a TechnicalIndicator and digs out its unparametrized type before running the original ismultiinput method.

TechnicalIndicatorCharts.ismultioutput — Method

This is a wrapper around OnlineTechnicalIndicators.ismultioutput that takes any instance of a TechnicalIndicator and digs out its unparametrized type before running the original ismultioutput method.

TechnicalIndicatorCharts.make_panel — Method

make_panel(chart::LWCChart)

Wrap a single LWCChart in a panel.

TechnicalIndicatorCharts.make_panel — Method

make_panel(plots::Vector)

Wrap a Vector of LWCCharts in a panel.

TechnicalIndicatorCharts.merge_candle! — Method

merge_candle!(last_candle, c)

If last candle is not provided, construct a new candle with the given OHLCV data. If last candle is provided, mutate lastcandle such that it's HLCV are updated. It's assumed that lastcandle and c have the same timestamp.

TechnicalIndicatorCharts.push_new_candle! — Method

This is meant to be called on timeframe boundaries to onto the chart's dataframe. It also does indicator calculation at this time.

TechnicalIndicatorCharts.update_last_candle! — Method

This is for internal housekeeping inside chart.candle. This happens when we're away from a chart.tf boundary. This doesn't go into a DataFrame.

TechnicalIndicatorCharts.visualize — Method

visualize(any::Any)

This is a visualize method that's a catch-all for indicators that haven't had a visualize method made for them yet. For now, it returns an empty panel.

TechnicalIndicatorCharts.visualize — Method

visualize(chart::Chart)

Return an LWCLayout that visualizes all the components in chart appropriately.

TechnicalIndicatorCharts.visualize — Method

visualize(sma::SMA)

Return an lwc_line for visualizing an SMA indicator.